+++ python-venv/cmstools/lib/common/bos_defs.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2024 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2022, 2024-2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -21,32 +21,13 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
-# Package info for barebones image boot test
 
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
+"""
+BOS URLs
+"""
 
-[tool.setuptools.packages]
-find = {}
+from .api import API_BASE_URL
 
-[project]
-name = "barebones_image_test"
-version = "@RPM_VERSION@"
-authors = [
-  { name="HPE Development LP", email="sps@cray.com" },
-]
-description = "Barebones image boot test"
-requires-python = ">=3.10"
-classifiers = [
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "License :: OSI Approved :: MIT License",
-    "Topic :: System :: Systems Administration",
-]
-
-[project.scripts]
-barebones_image_test = "cmstools.test.barebones_image_test.__main__:main"
-
-[project.urls]
-Homepage = "https://github.com/Cray-HPE/cms-tools"
+BOS_URL = f"{API_BASE_URL}/bos/v2"
+BOS_SESSIONS_URL = f"{BOS_URL}/sessions"
+BOS_TEMPLATES_URL = f"{BOS_URL}/sessiontemplates"
