+++ python-venv/cmstools/lib/common/defs.py
