+++ python-venv/cmstools/lib/common/s3/defs.py
